+++ web/app/page.tsx
@@ -1,5 +1,0 @@
-import { DailyLogApp } from "@/components/log/daily-log-app";
-
-export default function Home() {
-  return <DailyLogApp />;
-}